requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-documentai
google-api-core
python-dotenv
PyPDF2
openpyxl
reportlab
docx2pdf
python-docx
python-pptx
Pillow
aiofiles
pydantic
openai
motor
pymongo
langdetect
langchain
scikit-learn
numpy